move sample
